collector/requirements.txt
httpx==0.27.0
redis==5.0.4
python-dotenv==1.0.1
tenacity==8.2.3